ndk: Add mod media_error
Moved media::{construct, MediaErrorResult, NdkMediaError} to separate
module as midi.rs is also dependent on them
